+++ application/src/scenes/MenuScene.cpp
@@ -24,15 +24,15 @@ void MenuScene::setup() {
     // set background music
     playBGM("easy-cheesy-bgm");
 
-    auto _titleText = _registry.create();
-    _registry.emplace<TextComponent>(_titleText, "Sokoban", "JetBrainsMono-Regular", 32, 255, 255, 255, 255);
-    _registry.emplace<TransformComponent>(_titleText, 356, 194, 100, 50);
+    auto _titleText = _ecs.registry.create();
+    _ecs.registry.emplace<TextComponent>(_titleText, "Sokoban", "JetBrainsMono-Regular", 32, 255, 255, 255, 255);
+    _ecs.registry.emplace<TransformComponent>(_titleText, 356, 194, 100, 50);
 
-    auto _playButton = _registry.create();
-    _registry.emplace<TransformComponent>(_playButton, 339, 258, 150, 50);
-    _registry.emplace<TextComponent>(_playButton, "Play", "Kenney_Future", 32, Color{255, 255, 255, 255});
-    _registry.emplace<SpriteComponent>(_playButton, "button_rectangle_depth_flat", 0, 0, 384, 128);
-    _registry.emplace<ButtonComponent>(_playButton, [this]() {
+    auto _playButton = _ecs.registry.create();
+    _ecs.registry.emplace<TransformComponent>(_playButton, 339, 258, 150, 50);
+    _ecs.registry.emplace<TextComponent>(_playButton, "Play", "Kenney_Future", 32, Color{255, 255, 255, 255});
+    _ecs.registry.emplace<SpriteComponent>(_playButton, "button_rectangle_depth_flat", 0, 0, 384, 128);
+    _ecs.registry.emplace<ButtonComponent>(_playButton, [this]() {
         LOG_INFO("Changing scene to PlayScene");
         playSFX("click-a");
         changeScene("PlayScene");